+++ engine/test/unit/server/test_web.py
@@ -19,49 +19,55 @@ async def cleanup_test_server():
 def test_port_path():
     args = ['--port=8080', '--path=/tmp/test', '--config-files=engine.json,test.json']
     result = parse_args(args)
-    assert result == (None, 8080, '/tmp/test', False, ['engine.json', 'test.json'], None, [])
+    assert result == (None, 8080, '/tmp/test', False, ['engine.json', 'test.json'], None, [], [])
 
 
 def test_stream_start():
     args = ['--start-streams', '--config-files=test.json']
     result = parse_args(args)
-    assert result == (None, 8020, None, True, ['test.json'], None, [])
+    assert result == (None, 8020, None, True, ['test.json'], None, [], [])
 
 
 def test_port_only():
     args = ['--port=8020', '--config-files=test.json']
     result = parse_args(args)
-    assert result == (None, 8020, None, False, ['test.json'], None, [])
+    assert result == (None, 8020, None, False, ['test.json'], None, [], [])
 
 
 def test_path_only():
     args = ['--path=/tmp/test', '--config-files=test.json']
     result = parse_args(args)
-    assert result == (None, None, '/tmp/test', False, ['test.json'], None, [])
+    assert result == (None, None, '/tmp/test', False, ['test.json'], None, [], [])
 
 
 def test_rest_config_only():
     args = ['--config-files=test.json']
     result = parse_args(args)
-    assert result == (None, 8020, None, False, ['test.json'], None, [])
+    assert result == (None, 8020, None, False, ['test.json'], None, [], [])
 
 
 def test_rest_config_with_host():
     args = ['--host=127.0.0.1', '--config-files=test.json']
     result = parse_args(args)
-    assert result == ('127.0.0.1', 8020, None, False, ['test.json'], None, [])
+    assert result == ('127.0.0.1', 8020, None, False, ['test.json'], None, [], [])
 
 
 def test_config_with_api_file():
     args = ['--config-files=test.json', '--api-file=openapi.json']
     result = parse_args(args)
-    assert result == (None, 8020, None, False, ['test.json'], 'openapi.json', [])
+    assert result == (None, 8020, None, False, ['test.json'], 'openapi.json', [], [])
+
+
+def test_config_with_api_auto():
+    args = ['--config-files=test.json', '--api-auto=0.18;Title;Description']
+    result = parse_args(args)
+    assert result == (None, 8020, None, False, ['test.json'], None, ['0.18', 'Title', 'Description'], [])
 
 
 def test_config_with_groups():
     args = ['--config-files=test.json', '--api-file=openapi.json', '--enabled-groups=g1,g2']
     result = parse_args(args)
-    assert result == (None, 8020, None, False, ['test.json'], 'openapi.json', ['g1', 'g2'])
+    assert result == (None, 8020, None, False, ['test.json'], 'openapi.json', [], ['g1', 'g2'])
 
 
 class MockHooks:
@@ -88,8 +94,13 @@ async def _stream_startup_hook(*args, **kwargs):
     )
 
 
+@pytest.mark.parametrize("api_file,api_auto", [
+    ('test_api_file.json', []),
+    (None, ['0.18', 'Title', 'Description']),
+    (None, ['0.18']),
+    (None, None)])
 @pytest.mark.asyncio
-async def test_server_initialization(monkeypatch):
+async def test_server_initialization(monkeypatch, api_file, api_auto):
     async def _shutdown(*args, **kwargs):
         await asyncio.sleep(1)
         raise GracefulExit
@@ -97,7 +108,8 @@ async def test_server_initialization(monkeypatch):
     def _serve():
         web.init_web_server(
             config_files=['test_server_file.json', 'test_app_file.json', 'test_app_file2.json'],
-            api_file='test_api_file.json',
+            api_file=api_file,
+            api_auto=api_auto,
             enabled_groups=[],
             start_streams=True
         )
@@ -134,9 +146,14 @@ async def test_server_initialization(monkeypatch):
         assert len(MockHooks._stream_startup_hook_calls) == 2
 
         assert _load_engine_config.call_args[0] == ('test_server_file.json',)
-        assert _load_api_file.call_args[0] == ('test_api_file.json',)
-        assert _register_server_config.call_count == 1
+        if api_file:
+            assert _load_api_file.call_args[0] == ('test_api_file.json',)
+        assert _register_server_config.call_count == (1 if api_file or api_auto else 0)
         assert _load_app_config.call_args_list[0][0] == ('test_app_file.json',)
         assert _load_app_config.call_args_list[1][0] == ('test_app_file2.json',)
         assert _register_apps.call_count == 1
         assert _enable_swagger.call_count == 1
+
+        MockHooks._stream_startup_hook_calls = []
+        MockHooks._server_startup_hook_calls = []
+        MockHooks._app_startup_hook_calls = []